Backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
pydantic==1.10.11
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyMySQL==1.0.3
python-jose==3.3.0
email-validator==1.3.1
pytest==7.4.0
httpx==0.24.1
websockets==11.0.3
python-multipart==0.0.6
reportlab==4.0.4
pillow==10.0.0
